unicode: overhaul the way unicode codepoints are combined
Not all unicode codepoints that should be combined into previous
codepoints to create new graphemes are considered zero width.
Specifically, regional indicators by themselves are valid graphemes, but
when combined, they create various regional graphemes.

Additionally, there are a few emoji modifiers (for skin tone) that are
colors by themselves, but modify the previous grapheme if it is
combinable and _become_ zero width if so.

ZWJ (the zero-width joiner) is also not itself considered a combining
character, but could possibly cause two non-zero-width codepoints to all
be combined, and may increase the grapheme's width from one to two. VS16
can also modify a grapheme to force it to be rendered as an emoji, which
can increase its width from one to two.

All of this considering, the way mosh combines unicode codepoints into
cells needed some work. Now, it'll always check if it's possible to
combine with the previous cell and do it if it makes sense. This largely
follows tmux's general way of looking for combining opportunities.

Still a TODO is handling Hangul Jamo Korean combining which has its own
rules.